+++ BuddyBot/BuddyBot.csproj
@@ -129,6 +129,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="App_Start\WebApiConfig.cs" />
+    <Compile Include="Dialogs\Builders\RichCardAttachmentBuilder.cs" />
     <Compile Include="Services\Contracts\IConversationService.cs" />
     <Compile Include="Services\Contracts\IWeatherService.cs" />
     <Compile Include="Dialogs\Builders\DialogBuilder.cs" />
